# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Advancia Smart Wallets API
  version: "1.0.0"
  description: |
    REST API for the Advancia Smart Wallets platform — a fintech super-app
    with embedded smart wallets (ERC-4337), payments, healthcare, banking,
    loyalty, subscriptions, and admin management.

    ## Authentication
    Most endpoints require a JWT session token obtained from `/api/auth/*`.
    Pass it as a cookie (`session`) or `Authorization: Bearer <token>` header.

    Admin endpoints require a separate admin session via `/api/admin/login`.

    ## Versioning
    All endpoints are available under both `/api/*` and `/api/v1/*`.
    The `X-API-Version` response header indicates the active version.

    ## Rate Limiting
    Auth endpoints (`/api/auth/*`) are rate-limited to 20 requests/minute per IP.
    Exceeding the limit returns `429 Too Many Requests` with a `Retry-After` header.
  contact:
    name: Advancia Dev Team
    email: security@advancia.health
  license:
    name: Proprietary
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://advancia-healthcare1.vercel.app
    description: Production (Vercel)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
    cronAuth:
      type: http
      scheme: bearer
      description: CRON_SECRET bearer token for cron endpoints
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: "Invalid request body"
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Unauthorized"
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    TooManyRequests:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit window resets
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: "Too many requests. Please try again later."
...
security:
  - bearerAuth: []
  - cookieAuth: []

# ═══════════════════════════════════════════════════════════════
#  PATHS — grouped by tag / domain
# ═══════════════════════════════════════════════════════════════
paths:
...
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register via wallet signature
      security: []
      responses:
        "201": { description: "User created" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/auth/email/register:
    post:
      tags: [Auth]
      summary: Register with email and password
      security: []
      responses:
        "201": { description: "User created" }
        "400": { $ref: "#/components/responses/BadRequest" }
  /api/auth/email/login:
    post:
      tags: [Auth]
      summary: Login with email and password
      security: []
      responses:
        "200": { description: "Session created" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "429": { $ref: "#/components/responses/TooManyRequests" }
...
  /api/auth/pin:
    post:
      tags: [Auth]
      summary: Set or verify PIN
      responses:
        "200": { description: "PIN accepted" }
        "400": { $ref: "#/components/responses/BadRequest" }
...
  /api/admin/login:
    post:
      tags: [Admin]
      summary: Admin login
      security: []
      responses:
        "200": { description: "Admin session created" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    delete:
      tags: [Admin]
      summary: Admin logout
      responses:
        "200": { description: "Admin session destroyed" }
...